requirements.txt
gspread
oauth2client
duckdb
airflow-provider-duckdb
astro-sdk-python[duckdb]
